requirements.txt
# Requirements for Obsidian Image Processor
watchdog>=3.0.0
Pillow>=9.0.0